requirements.txt
numpy=1.15.2
Pillow=5.1.0
